Extract the Converter7300 class.
