+++ FbxSharp.csproj
@@ -148,6 +148,7 @@
     <Compile Include="TokenizationException.cs" />
     <Compile Include="ParseException.cs" />
     <Compile Include="ConversionException.cs" />
+    <Compile Include="Converter7300.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
